Procedural Progress
Great progress stairs spawning paths, needs better connection logic. Major thing to do is smoothing the map. Then bridges need to be added. Then 'Second level stair check.
Also need to implement a system to keep track of the node height and status. Once these are done we can add in the GUI Buttons and we are set for gameplay.